changed file from js to tsx, changed function to parameterised function
