+++ tsconfig.json
@@ -15,6 +15,6 @@
     "jsx": "preserve",
     "incremental": true
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "features/home/components/News.tsx", "features/home/components/ClubsOverview.tsx"],
   "exclude": ["node_modules"]
 }
